Separate tests into new file
